Split out check your booking tests
